@@ -9,60 +9,53 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "test-ci": "ng test --watch=false --browsers=ChromeHeadless",
-    "build-prod": "ng build --prod --build-optimizer",
-    "yalc-add-modules": "yalc add @knora/action && yalc add @knora/authentication && yalc add @knora/core && yalc add @knora/search && yalc add @knora/viewer"
+    "build-prod": "ng build --prod --build-optimizer"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.3",
-    "@angular/cdk": "^8.0.1",
-    "@angular/common": "^8.0.3",
-    "@angular/compiler": "^8.0.3",
-    "@angular/core": "^8.0.3",
-    "@angular/flex-layout": "^8.0.0-beta.26",
-    "@angular/forms": "^8.0.3",
-    "@angular/material": "^8.1.4",
-    "@angular/platform-browser": "^8.0.3",
-    "@angular/platform-browser-dynamic": "^8.0.3",
-    "@angular/router": "^8.0.3",
-    "@knora/action": "^9.1.0",
-    "@knora/core": "^9.1.0",
-    "@knora/search": "^9.1.0",
-    "@knora/viewer": "^9.1.0",
-    "core-js": "^2.5.4",
-    "hammerjs": "^2.0.8",
+    "@angular/animations": "^9.1.12",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.12",
+    "@angular/compiler": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/flex-layout": "^9.0.0-beta.31",
+    "@angular/forms": "^9.1.12",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "^9.1.12",
+    "@angular/platform-browser-dynamic": "^9.1.12",
+    "@angular/router": "^9.1.12",
+    "@dasch-swiss/dsp-js": "1.0.0-rc.10",
+    "@dasch-swiss/dsp-ui": "1.0.0-rc.5",
+    "core-js": "^3.6.5",
     "jdnconvertiblecalendar": "^0.0.5",
-    "jdnconvertiblecalendardateadapter": "^0.0.10",
-    "json2typescript": "^1.0.6",
-    "jsonld": "^1.0.4",
-    "ngx-infinite-scroll": "^8.0.0",
-    "node-sass": "^4.9.3",
+    "jdnconvertiblecalendardateadapter": "^0.0.13",
+    "ngx-color-picker": "^9.1.0",
     "openseadragon": "^2.4.0",
     "rxjs": "^6.5.2",
     "svg-overlay": "github:openseadragon/svg-overlay",
     "ts-md5": "^1.2.4",
-    "tslib": "^1.9.0",
+    "tslib": "^1.10.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.2",
-    "@angular/cli": "~8.3.2",
-    "@angular/compiler-cli": "^8.0.3",
-    "@angular/language-service": "^8.0.3",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "~0.901.12",
+    "@angular/cli": "~9.1.12",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/language-service": "^9.1.12",
+    "@types/jasmine": "~3.4.4",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.1",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
+    "karma": "^5.1.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~3.3.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "^7.0.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.5.3"
+    "typescript": "~3.8.3"
   }
 }
