@@ -1,38 +1,48 @@
-import { TestBed, async } from '@angular/core/testing';
+import { async, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
-import { KuiSearchModule } from '@knora/search';
 import { MaterialModule } from './material-module';
-import { KuiCoreConfig, KuiCoreConfigToken } from '@knora/core';
+import { Component } from '@angular/core';
+
+/**
+ * Test host component to simulate dsp-search-panel component.
+ */
+@Component({
+    selector: 'dsp-search-panel',
+    template: ''
+})
+class DspSearchPanelComponent {
+
+}
 
 describe('AppComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        KuiSearchModule,
-        MaterialModule,
-        RouterTestingModule
-      ],
-      declarations: [
-        AppComponent
-      ],
-      providers: [{ provide: KuiCoreConfigToken, useValue: KuiCoreConfig }]
-    }).compileComponents();
-  }));
-  it('should create the app', async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app).toBeTruthy();
-  }));
-  it(`should have as title 'app'`, async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('app');
-  }));
-  /* it('should render title in a h1 tag', async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Welcome to beol!');
-  })); */
+
+    beforeEach(async(() => {
+
+        TestBed.configureTestingModule({
+            imports: [
+                MaterialModule,
+                RouterTestingModule
+            ],
+            declarations: [
+                AppComponent,
+                DspSearchPanelComponent
+            ],
+            providers: []
+        }).compileComponents();
+
+
+    }));
+
+    it('should create the app', async(() => {
+        const fixture = TestBed.createComponent(AppComponent);
+        const app = fixture.debugElement.componentInstance;
+        expect(app).toBeTruthy();
+    }));
+
+    it(`should have as title 'app'`, async(() => {
+        const fixture = TestBed.createComponent(AppComponent);
+        const app = fixture.debugElement.componentInstance;
+        expect(app.title).toEqual('app');
+    }));
 });
