@@ -1,11 +1,19 @@
 // angular modules
 import { BrowserModule } from '@angular/platform-browser';
 import { APP_INITIALIZER, NgModule } from '@angular/core';
-// @knora modules
-import { KuiCoreConfigToken, KuiCoreModule } from '@knora/core';
-import { KuiActionModule } from '@knora/action';
-import { KuiSearchModule } from '@knora/search';
-import { KuiViewerModule } from '@knora/viewer';
+import { environment } from '../environments/environment';
+// @dsp-js library
+import { KnoraApiConnection } from '@dasch-swiss/dsp-js';
+// @dsp-ui library
+import {
+    AppInitService,
+    DspActionModule,
+    DspApiConfigToken,
+    DspApiConnectionToken,
+    DspCoreModule,
+    DspSearchModule,
+    DspViewerModule
+} from '@dasch-swiss/dsp-ui';
 // modules from @angular/material and @angular/flex-layout
 import { MaterialModule } from './material-module';
 import { FlexLayoutModule } from '@angular/flex-layout';
@@ -37,8 +45,6 @@ import { CorrespondenceComponent } from './correspondence/correspondence.compone
 import { ContactComponent } from './contact/contact.component';
 
 import { ReactiveFormsModule } from '@angular/forms';
-// import { RECAPTCHA_SETTINGS, RecaptchaSettings } from 'ng-recaptcha';
-import { InfiniteScrollModule } from 'ngx-infinite-scroll';
 
 import { PageComponent } from './resource/page/page.component';
 import { ReadTextValueComponent } from './properties/read-text-value/read-text-value.component';
@@ -51,19 +57,12 @@ import { LeibnizPortalDirective } from './directives/leibniz-portal.directive';
 import { HanCatalogueDirective } from './directives/han-catalogue.directive';
 import { BebbRouteComponent } from './bebb-route/bebb-route.component';
 import { MAT_DIALOG_DEFAULT_OPTIONS } from '@angular/material/dialog';
-import { AppInitService } from './app-init.service';
 
 import { TranscriptionComponent } from './resource/transcription/transcription.component';
 import { ManuscriptEntryComponent } from './resource/manuscript-entry/manuscript-entry.component';
 import { TeiLinkDirective } from './directives/tei-link.directive';
 import { CommentComponent } from './resource/comment/comment.component';
-
-
-export function initializeApp(appInitService: AppInitService) {
-    return (): Promise<any> => {
-        return appInitService.Init();
-    };
-}
+import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
     declarations: [
@@ -106,21 +105,33 @@ export function initializeApp(appInitService: AppInitService) {
         AppRouting,
         BrowserModule,
         FlexLayoutModule,
-        InfiniteScrollModule,
-        KuiCoreModule,
-        KuiActionModule,
-        KuiSearchModule,
-        KuiViewerModule,
+        DspCoreModule,
+        DspViewerModule,
+        DspActionModule,
+        DspSearchModule,
         MaterialModule,
-        ReactiveFormsModule
+        ReactiveFormsModule,
+        HttpClientModule
     ],
     providers: [
-        AppInitService,
         {
-            provide: APP_INITIALIZER, useFactory: initializeApp, deps: [AppInitService], multi: true
+            provide: APP_INITIALIZER,
+            useFactory: (appInitService: AppInitService) =>
+                (): Promise<void> => {
+                    return appInitService.Init('config', environment);
+                },
+            deps: [AppInitService],
+            multi: true
         },
         {
-            provide: KuiCoreConfigToken, useFactory: () => AppInitService.coreConfig
+            provide: DspApiConfigToken,
+            useFactory: (appInitService: AppInitService) => appInitService.dspApiConfig,
+            deps: [AppInitService]
+        },
+        {
+            provide: DspApiConnectionToken,
+            useFactory: (appInitService: AppInitService) => new KnoraApiConnection(appInitService.dspApiConfig),
+            deps: [AppInitService]
         },
         {
             provide: MAT_DIALOG_DEFAULT_OPTIONS,
