@@ -1,41 +1,30 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { HttpClientModule } from '@angular/common/http';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-
 import { MaterialModule } from '../material-module';
 import { CorrespondenceComponent } from './correspondence.component';
-import { KuiCoreConfig, KuiCoreConfigToken } from '@knora/core';
-import { AppInitService } from '../app-init.service';
+import { BeolService } from '../services/beol.service';
 
 describe('CorrespondenceComponent', () => {
     let component: CorrespondenceComponent;
     let fixture: ComponentFixture<CorrespondenceComponent>;
 
-    let appInitService: AppInitService;
+    const beolServiceSpy = jasmine.createSpyObj('BeolService', ['searchForCorrespondence', 'searchForNewtonCorrespondence', 'searchForLeibnizCorrespondence']); // see https://angular.io/guide/testing#angular-testbed
 
     beforeEach(async(() => {
-
-        const appInitServiceSpy = jasmine.createSpyObj('AppInitService', ['getSettings']);
 
         TestBed.configureTestingModule({
             imports: [
                 RouterTestingModule,
-                MaterialModule,
-                HttpClientModule,
-                HttpClientTestingModule
+                MaterialModule
             ],
             declarations: [CorrespondenceComponent],
             providers: [
-                {provide: KuiCoreConfigToken, useValue: KuiCoreConfig},
-                {provide: AppInitService, useValue: appInitServiceSpy}
+                { provide: BeolService, useValue: beolServiceSpy }
             ]
         })
             .compileComponents();
 
-        appInitServiceSpy.getSettings.and.returnValue({ontologyIRI: 'http://0.0.0.0:3333'});
 
-        appInitService = TestBed.get(AppInitService);
     }));
 
     beforeEach(() => {
