@@ -1,40 +1,33 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ReadTextValueComponent } from './read-text-value.component';
-import { Component, DebugElement, OnInit, ViewChild } from '@angular/core';
-import {
-    KuiCoreConfig, KuiCoreConfigToken,
-    OntologyInformation,
-    ReadTextValue,
-    ReadTextValueAsHtml,
-    ReadTextValueAsString,
-    ReadTextValueAsXml,
-    ResourceClass,
-    ResourceClasses,
-    ResourceClassIrisForOntology
-} from '@knora/core';
-import { KuiViewerModule, TextValueAsStringComponent, TextValueAsXmlComponent } from '@knora/viewer';
-import { ReadTextValueAsHtmlComponent } from '../read-text-value-as-html/read-text-value-as-html.component';
-import { MathJaxDirective } from '../../directives/mathjax.directive';
-import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { AppInitService } from '../../app-init.service';
+import { ReadTextValueComponent } from './read-text-value.component';
+import { ReadTextValueAsHtmlComponent } from '../read-text-value-as-html/read-text-value-as-html.component';
+import { MathJaxDirective } from '../../directives/mathjax.directive';
+import { Component, DebugElement, OnInit, ViewChild } from '@angular/core';
+import { DspViewerModule, TextValueAsStringComponent } from '@dasch-swiss/dsp-ui';
+import { By } from '@angular/platform-browser';
+import {
+    Constants,
+    ReadResource,
+    ReadTextValue,
+    ReadTextValueAsHtml,
+    ReadTextValueAsString
+} from '@dasch-swiss/dsp-js';
+import { BeolService } from '../../services/beol.service';
 
 describe('ReadTextValueComponent', () => {
     let testHostComponent: TestHostComponent;
     let testHostFixture: ComponentFixture<TestHostComponent>;
-
-    let appInitService: AppInitService;
 
     beforeEach(async(() => {
         const appInitServiceSpy = jasmine.createSpyObj('AppInitService', ['getSettings']);
 
         TestBed.configureTestingModule({
             imports: [
-                KuiViewerModule,
                 RouterTestingModule,
-                MatSnackBarModule
+                MatSnackBarModule,
+                DspViewerModule
             ],
             declarations: [
                 ReadTextValueComponent,
@@ -43,15 +36,15 @@ describe('ReadTextValueComponent', () => {
                 TestHostComponent
             ],
             providers: [
-                { provide: KuiCoreConfigToken, useValue: KuiCoreConfig },
-                { provide: AppInitService, useValue: appInitServiceSpy }
+                {
+                    provide: BeolService, useValue: {} // mock BeolService because it has its own deps
+                }
             ]
         })
             .compileComponents();
 
         appInitServiceSpy.getSettings.and.returnValue({ ontologyIRI: 'http://0.0.0.0:3333' });
 
-        appInitService = TestBed.get(AppInitService);
     }));
 
     beforeEach(() => {
@@ -66,11 +59,10 @@ describe('ReadTextValueComponent', () => {
         expect(testHostComponent.textValueComponent).toBeTruthy();
 
         expect(testHostComponent.textValueComponent.valueObject.id).toEqual('http://rdfh.ch/0802/V/values/Q');
-        expect(testHostComponent.textValueComponent.valueObject.propIri).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasName');
-        expect(testHostComponent.textValueComponent.valueObject.getClassName()).toEqual('ReadTextValueAsString');
+        expect(testHostComponent.textValueComponent.valueObject.property).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasName');
 
         expect(testHostComponent.textValueComponent.bindEvents).toBeFalsy();
-        expect(testHostComponent.textValueComponent.ontologyInfo).toEqual(ontoInfo);
+
 
     });
 
@@ -86,47 +78,51 @@ describe('ReadTextValueComponent', () => {
 
         const textValComp: TextValueAsStringComponent = textValDe.componentInstance as TextValueAsStringComponent;
 
-        expect(textValComp.valueObject.id).toEqual('http://rdfh.ch/0802/V/values/Q');
+        expect(textValComp.displayValue.id).toEqual('http://rdfh.ch/0802/V/values/Q');
 
-        expect(textValComp.valueObject.propIri).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasName');
+        expect(textValComp.displayValue.property).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasName');
 
-        expect(textValComp.valueObject.str).toEqual('test string');
-
-    });
-
-    it('should have created a TextValueAsXmlComponent with the correct value', () => {
-
-        testHostComponent.textVal =
-            new ReadTextValueAsXml('http://rdfh.ch/0802/V/values/Y',
-                'http://0.0.0.0/ontology/0801/beol/v2#hasXml',
-                '<div>test</div>', 'http://rdfh.ch/0802/mappings/test');
-
-        testHostFixture.detectChanges();
-
-        const hostCompDe = testHostFixture.debugElement;
-
-        const textValDe: DebugElement = hostCompDe.query(By.directive(TextValueAsXmlComponent));
-
-        expect(textValDe).toBeTruthy();
-
-        expect(textValDe.componentInstance).toBeTruthy();
-
-        const textValComp: TextValueAsXmlComponent = textValDe.componentInstance as TextValueAsXmlComponent;
-
-        expect(textValComp.valueObject.id).toEqual('http://rdfh.ch/0802/V/values/Y');
-
-        expect(textValComp.valueObject.propIri).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasXml');
-
-        expect(textValComp.valueObject.xml).toEqual('<div>test</div>');
+        expect(textValComp.displayValue.text).toEqual('test string');
 
     });
+
+    /*it('should have created a TextValueAsXmlComponent with the correct value', () => {
+
+      testHostComponent.textVal =
+          new ReadTextValueAsXml('http://rdfh.ch/0802/V/values/Y',
+              'http://0.0.0.0/ontology/0801/beol/v2#hasXml',
+              '<div>test</div>', 'http://rdfh.ch/0802/mappings/test');
+
+      testHostFixture.detectChanges();
+
+      const hostCompDe = testHostFixture.debugElement;
+
+      const textValDe: DebugElement = hostCompDe.query(By.directive(TextValueAsXmlComponent));
+
+      expect(textValDe).toBeTruthy();
+
+      expect(textValDe.componentInstance).toBeTruthy();
+
+      const textValComp: TextValueAsXmlComponent = textValDe.componentInstance as TextValueAsXmlComponent;
+
+      expect(textValComp.valueObject.id).toEqual('http://rdfh.ch/0802/V/values/Y');
+
+      expect(textValComp.valueObject.propIri).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasXml');
+
+      expect(textValComp.valueObject.xml).toEqual('<div>test</div>');
+
+    });*/
 
     it('should have created a ReadTextAsHtmlComponent with the correct value', () => {
 
-        testHostComponent.textVal =
-            new ReadTextValueAsHtml('http://rdfh.ch/0802/V/values/Z',
-                'http://0.0.0.0/ontology/0801/beol/v2#hasText',
-                '<span>test</span>', {});
+        const htmlVal = new ReadTextValueAsHtml();
+        htmlVal.id = 'http://rdfh.ch/0802/V/values/Z';
+        htmlVal.property = 'http://0.0.0.0/ontology/0801/beol/v2#hasText';
+        htmlVal.html = '<div>test</div>';
+        htmlVal.strval = htmlVal.html;
+        htmlVal.type = Constants.TextValue;
+
+        testHostComponent.textVal = htmlVal;
 
         testHostComponent.bindEvents = true;
 
@@ -144,13 +140,12 @@ describe('ReadTextValueComponent', () => {
 
         expect(textValComp.valueObject.id).toEqual('http://rdfh.ch/0802/V/values/Z');
 
-        expect(textValComp.valueObject.propIri).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasText');
+        expect(textValComp.valueObject.property).toEqual('http://0.0.0.0/ontology/0801/beol/v2#hasText');
 
-        expect(textValComp.valueObject.html).toEqual('<span>test</span>');
+        expect(textValComp.valueObject.html).toEqual('<div>test</div>');
 
         expect(textValComp.bindEvents).toBeTruthy();
 
-        expect(textValComp.ontologyInfo).toEqual(ontoInfo);
     });
 });
 
@@ -162,7 +157,7 @@ describe('ReadTextValueComponent', () => {
         <read-text-value #textValComp
                          [valueObject]="textVal"
                          [bindEvents]="bindEvents"
-                         [ontologyInfo]="ontologyInfo">
+                         [resource]="res">
         </read-text-value>`
 })
 class TestHostComponent implements OnInit {
@@ -171,40 +166,21 @@ class TestHostComponent implements OnInit {
 
     textVal: ReadTextValue;
     bindEvents = false;
-    ontologyInfo;
+    res: ReadResource;
 
-    constructor () {
+    constructor() {
     }
 
     ngOnInit() {
-        this.textVal =
-            new ReadTextValueAsString(
-                'http://rdfh.ch/0802/V/values/Q',
-                'http://0.0.0.0/ontology/0801/beol/v2#hasName',
-                'test string');
+        const textVal = new ReadTextValueAsString();
+        textVal.id = 'http://rdfh.ch/0802/V/values/Q';
+        textVal.property = 'http://0.0.0.0/ontology/0801/beol/v2#hasName';
+        textVal.text = 'test string';
+        textVal.type = Constants.TextValue;
 
-        this.ontologyInfo = ontoInfo;
+        this.textVal = textVal;
+
+        this.res = new ReadResource();
+
     }
 }
-
-const resClassesForOnto: ResourceClassIrisForOntology = {
-    'http://0.0.0.0:3333/ontology/0803/incunabula/v2': [
-        'http://0.0.0.0:3333/ontology/0803/incunabula/v2#book'
-    ]
-};
-
-const resClasses: ResourceClasses = {
-    'http://0.0.0.0:3333/ontology/0803/incunabula/v2#book':
-        new ResourceClass(
-            'http://0.0.0.0:3333/ontology/0803/incunabula/v2#book',
-            'book.png',
-            'A book.',
-            'book',
-            [],
-            []
-        )
-};
-
-const ontoInfo = new OntologyInformation(resClassesForOnto, resClasses, {});
-
-
