@@ -1,56 +1,64 @@
+import { Directive, Inject, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import {
-    IncomingService,
-    KnoraConstants,
-    OntologyCacheService,
-    OntologyInformation,
+    Constants,
+    KnoraApiConnection,
     ReadLinkValue,
-    ReadPropertyItem,
     ReadResource,
-    ReadResourcesSequence,
+    ReadResourceSequence,
     ReadStillImageFileValue,
-    ResourceService,
-    StillImageRepresentation,
-    Utils
-} from '@knora/core';
-
-import { ImageRegion, StillImageComponent } from '@knora/viewer';
-
+    ReadValue
+} from '@dasch-swiss/dsp-js';
 import { Subscription } from 'rxjs';
-
-import { OnDestroy, OnInit, ViewChild } from '@angular/core';
-
-import { BeolService } from '../services/beol.service';
+import { DspApiConnectionToken, Region, StillImageComponent, StillImageRepresentation } from '@dasch-swiss/dsp-ui';
 import { ActivatedRoute, ParamMap } from '@angular/router';
+import { IncomingService } from '../services/incoming.service';
+import { BeolService } from '../services/beol.service';
+
+export class BeolCompoundResource {
+
+    readResource: ReadResource;
+
+    incomingRegions: ReadResource[] = [];
+
+    incomingStillImageRepresentations: ReadResource[] = [];
+
+    stillImageRepresentationsToDisplay: StillImageRepresentation[] = [];
+
+    constructor(resource: ReadResource) {
+
+        this.readResource = resource;
+    }
+}
 
 export interface PropIriToNameMapping {
     [index: string]: string;
 }
 
 export interface PropertyValues {
-    [index: string]: ReadPropertyItem[];
+    [index: string]: ReadValue[];
 }
 
+
+@Directive()
 export abstract class BeolResource implements OnInit, OnDestroy {
 
     abstract iri: string;
-    abstract resource: ReadResource;
-    abstract ontologyInfo: OntologyInformation;
-    abstract isLoading: boolean;
+    abstract resource: BeolCompoundResource;
+    abstract isLoading = true;
     abstract errorMessage: any;
     abstract incomingStillImageRepresentationCurrentOffset: number;
     abstract navigationSubscription: Subscription;
     protected params;
 
-    @ViewChild('OSDViewer', { static: false }) osdViewer: StillImageComponent;
+    @ViewChild('OSDViewer') osdViewer: StillImageComponent;
 
-    abstract KnoraConstants: KnoraConstants;
+    abstract dspConstants: Constants;
 
     abstract propIris: PropIriToNameMapping;
 
-    constructor (
+    constructor(
+        @Inject(DspApiConnectionToken) protected _dspApiConnection: KnoraApiConnection,
         protected _route: ActivatedRoute,
-        protected _resourceService: ResourceService,
-        protected _cacheService: OntologyCacheService,
         protected _incomingService: IncomingService,
         protected _beolService: BeolService) {
     }
@@ -59,29 +67,25 @@ export abstract class BeolResource implements OnInit, OnDestroy {
      * Creates a collection of [[StillImageRepresentation]] belonging to the given resource and assigns it to it.
      * Each [[StillImageRepresentation]] represents an image including regions.
      *
-     * @param {ReadResource} resource          The resource to get the images for.
-     * @returns {StillImageRepresentation[]}   A collection of images for the given resource.
+     * @param resource  The resource to get the images for.
+     * @returns A collection of images for the given resource.
      */
-    protected static collectImagesAndRegionsForResource(resource: ReadResource): void {
+    protected static collectImagesAndRegionsForResource(resource: BeolCompoundResource): void {
 
         const imgRepresentations: StillImageRepresentation[] = [];
 
-        if (resource.properties[KnoraConstants.hasStillImageFileValue] !== undefined) {
+        if (resource.readResource.properties[Constants.HasStillImageFileValue] !== undefined) {
             // TODO: check if resources is a StillImageRepresentation using the ontology responder (support for subclass relations required)
             // resource has StillImageFileValues that are directly attached to it (properties)
 
-            const fileValues: ReadStillImageFileValue[] = resource.properties[KnoraConstants.hasStillImageFileValue] as ReadStillImageFileValue[];
-            const imagesToDisplay: ReadStillImageFileValue[] = fileValues.filter((image) => {
-                return !image.isPreview;
-            });
+            const fileValues: ReadStillImageFileValue[] = resource.readResource.properties[Constants.HasStillImageFileValue] as ReadStillImageFileValue[];
 
+            for (const img of fileValues) {
 
-            for (const img of imagesToDisplay) {
-
-                const regions: ImageRegion[] = [];
+                const regions: Region[] = [];
                 for (const incomingRegion of resource.incomingRegions) {
 
-                    const region = new ImageRegion(incomingRegion);
+                    const region = new Region(incomingRegion);
 
                     regions.push(region);
 
@@ -98,25 +102,22 @@ export abstract class BeolResource implements OnInit, OnDestroy {
 
             const readStillImageFileValues: ReadStillImageFileValue[] = resource.incomingStillImageRepresentations.map(
                 (stillImageRes: ReadResource) => {
-                    const fileValues = stillImageRes.properties[KnoraConstants.hasStillImageFileValue] as ReadStillImageFileValue[];
+                    const fileValues = stillImageRes.properties[Constants.HasStillImageFileValue] as ReadStillImageFileValue[];
                     // TODO: check if resources is a StillImageRepresentation using the ontology responder (support for subclass relations required)
-                    const imagesToDisplay = fileValues.filter((image) => {
-                        return !image.isPreview;
-                    });
 
-                    return imagesToDisplay;
+                    return fileValues;
                 }
-            ).reduce(function (prev, curr) {
+            ).reduce((prev, curr) => {
                 // transform ReadStillImageFileValue[][] to ReadStillImageFileValue[]
                 return prev.concat(curr);
             });
 
             for (const img of readStillImageFileValues) {
 
-                const regions: ImageRegion[] = [];
+                const regions: Region[] = [];
                 for (const incomingRegion of resource.incomingRegions) {
 
-                    const region = new ImageRegion(incomingRegion);
+                    const region = new Region(incomingRegion);
                     regions.push(region);
 
                 }
@@ -146,11 +147,150 @@ export abstract class BeolResource implements OnInit, OnDestroy {
         return invertedMapping;
     }
 
+    /**
+     * Requests incoming resources for [[this.resource]].
+     * Incoming resources are: regions, StillImageRepresentations, and incoming links.
+     *
+     *
+     */
+    protected requestIncomingResources(): void {
+
+        // make sure that this.resource has been initialized correctly
+        if (this.resource === undefined) {
+            return;
+        }
+
+        // request incoming regions
+        if (this.resource.readResource.properties[Constants.HasStillImageFileValue]) {
+            // TODO: check if resources is a StillImageRepresentation using the ontology responder (support for subclass relations required)
+            // the resource is a StillImageRepresentation, check if there are regions pointing to it
+
+            this.getIncomingRegions(0);
+
+        } else {
+            // this resource is not a StillImageRepresentation
+            // check if there are StillImageRepresentations pointing to this resource
+
+            // this gets the first page of incoming StillImageRepresentations
+            // more pages may be requested by [[this.viewer]].
+            // TODO: for now, we begin with offset 0. This may have to be changed later (beginning somewhere in a collection)
+            this.getIncomingStillImageRepresentations(0);
+        }
+
+        // check for incoming links for the current resource
+        this.getIncomingLinks(0);
+
+    }
+
+    /**
+     * Gets the incoming regions for [[this.resource]].
+     *
+     * @param offset the offset to be used (needed for paging). First request uses an offset of 0.
+     */
+    protected getIncomingRegions(offset: number): void {
+        this._incomingService.getIncomingRegions(this.resource.readResource.id, offset).subscribe(
+            (regions: ReadResourceSequence) => {
+
+                // Append elements of regions.resources to resource.incoming
+                Array.prototype.push.apply(this.resource.incomingRegions, regions.resources);
+
+                // prepare regions to be displayed
+                // triggers ngOnChanges of StillImageComponent
+                BeolResource.collectImagesAndRegionsForResource(this.resource);
+
+            },
+            (error: any) => {
+                this.errorMessage = error;
+                this.isLoading = false;
+            }
+        );
+    }
+
+    /**
+     * Get StillImageRepresentations pointing to [[this.resource]].
+     * This method may have to called several times with an increasing offsetChange in order to get all available StillImageRepresentations.
+     *
+     * @param offset the offset to be used (needed for paging). First request uses an offset of 0.
+     * It takes the number of images returned as an argument.
+     */
+    protected getIncomingStillImageRepresentations(offset: number): void {
+        // make sure that this.resource has been initialized correctly
+        if (this.resource === undefined) {
+            return;
+        }
+
+        if (offset < 0) {
+            console.log(`offset of ${offset} is invalid`);
+            return;
+        }
+
+        this._incomingService.getStillImageRepresentationsForCompoundResource(this.resource.readResource.id, offset).subscribe(
+            (incomingImageRepresentations: ReadResourceSequence) => {
+
+                if (incomingImageRepresentations.resources.length > 0) {
+
+                    // set current offset
+                    this.incomingStillImageRepresentationCurrentOffset = offset;
+
+                    // TODO: implement prepending of StillImageRepresentations when moving to the left (getting previous pages)
+                    // TODO: append existing images to response and then assign response to `this.resource.incomingStillImageRepresentations`
+                    // TODO: maybe we have to support non consecutive arrays (sparse arrays)
+
+                    // append incomingImageRepresentations.resources to this.resource.incomingStillImageRepresentations
+                    Array.prototype.push.apply(this.resource.incomingStillImageRepresentations, incomingImageRepresentations.resources);
+
+                    // prepare attached image files to be displayed
+                    BeolResource.collectImagesAndRegionsForResource(this.resource);
+                }
+            },
+            (error: any) => {
+                this.errorMessage = error;
+                this.isLoading = false;
+            }
+        );
+
+    }
+
+    /**
+     * Get resources pointing to [[this.resource]] with properties other than knora-api:isPartOf and knora-api:isRegionOf.
+     *
+     * @param offset the offset to be used (needed for paging). First request uses an offset of 0.
+     * It takes the number of images returned as an argument.
+     */
+    protected getIncomingLinks(offset: number): void {
+
+        this._incomingService.getIncomingLinksForResource(this.resource.readResource.id, offset).subscribe(
+            (incomingResources: ReadResourceSequence) => {
+
+                // Append elements incomingResources to this.resource.incomingLinks
+                Array.prototype.push.apply(this.resource.readResource.incomingReferences, incomingResources.resources);
+            },
+            (error: any) => {
+                this.errorMessage = error;
+                this.isLoading = false;
+            }
+        );
+    }
+
+    /**
+     * The user clicked on an internal link.
+     *
+     * @param linkVal the value representing the referred resource.
+     */
+    protected resLinkClicked(linkVal: ReadLinkValue) {
+
+        const refResType = (linkVal.linkedResource !== undefined ? linkVal.linkedResource.type : '');
+
+        this._beolService.routeByResourceType(refResType, linkVal.linkedResourceIri);
+    }
+
     ngOnInit() {
         this.navigationSubscription = this._route.paramMap.subscribe((params: ParamMap) => {
             this.params = params;
             this.iri = params.get('id');
-            this.getResource(this.iri);
+            if (this.iri) {
+                this.getResource(this.iri);
+            }
         });
 
     }
@@ -176,11 +316,10 @@ export abstract class BeolResource implements OnInit, OnDestroy {
 
         const swapped = BeolResource.swap(this.propIris);
 
-        for (const key in this.resource.properties) {
-            if (this.resource.properties.hasOwnProperty(key)) {
-                for (const val of this.resource.properties[key]) {
-
-                    const name = swapped[val.propIri];
+        for (const key in this.resource.readResource.properties) {
+            if (this.resource.readResource.properties.hasOwnProperty(key)) {
+                for (const val of this.resource.readResource.properties[key]) {
+                    const name = swapped[val.property];
 
                     if (name !== undefined && Array.isArray(propClass[name])) {
                         propClass[name].push(val);
@@ -191,234 +330,37 @@ export abstract class BeolResource implements OnInit, OnDestroy {
     }
 
     /**
-     * The user clicked on an internal link.
-     *
-     * @param linkVal the value reprenting the referred resource.
-     */
-    protected resLinkClicked(linkVal: ReadLinkValue) {
-
-        const refResType = (linkVal.referredResource !== undefined ? linkVal.referredResource.type : '');
-
-        this._beolService.routeByResourceType(refResType, linkVal.referredResourceIri);
-    }
-
-    /**
      * Requests a resource.
      *
      * @param iri the Iri of the resource to be requested.
      */
     getResource(iri: string): void {
 
-        this._resourceService.getReadResource(iri)
+        this._dspApiConnection.v2.res.getResource(iri)
             .subscribe(
-                (result: ReadResourcesSequence) => {
+                (result: ReadResource) => {
 
-                    // make sure that exactly one resource is returned
-                    if (result.resources.length === 1) {
+                    // console.log(result)
 
-                        // initialize ontology information
-                        this.ontologyInfo = result.ontologyInformation;
+                    const res = new BeolCompoundResource(result);
 
-                        // prepare a possibly attached image file to be displayed
-                        BeolResource.collectImagesAndRegionsForResource(result.resources[0]);
+                    // prepare a possibly attached image file to be displayed
+                    BeolResource.collectImagesAndRegionsForResource(res);
 
-                        this.resource = result.resources[0];
+                    this.resource = res;
 
-                        this.initProps();
+                    this.initProps();
 
-                        this.isLoading = false;
+                    this.isLoading = false;
 
-                        this.requestIncomingResources();
-
-                    } else {
-                        // exactly one resource was expected, but resourceSeq.resources.length != 1
-                        this.errorMessage = `Exactly one resource was expected, but ${result.resources.length} resource(s) given.`;
-
-                    }
+                    this.requestIncomingResources();
 
                 },
                 (error: any) => {
-                    this.errorMessage = <any>error;
+                    this.errorMessage = error;
                     this.isLoading = false;
                 }
             );
     }
 
-    /**
-     * Requests incoming resources for [[this.resource]].
-     * Incoming resources are: regions, StillImageRepresentations, and incoming links.
-     *
-     **/
-    protected requestIncomingResources(): void {
-
-        // make sure that this.resource has been initialized correctly
-        if (this.resource === undefined) {
-            return;
-        }
-
-        // request incoming regions
-        if (this.resource.properties[KnoraConstants.hasStillImageFileValue]) {
-            // TODO: check if resources is a StillImageRepresentation using the ontology responder (support for subclass relations required)
-            // the resource is a StillImageRepresentation, check if there are regions pointing to it
-
-            this.getIncomingRegions(0);
-
-        } else {
-            // this resource is not a StillImageRepresentation
-            // check if there are StillImageRepresentations pointing to this resource
-
-            // this gets the first page of incoming StillImageRepresentations
-            // more pages may be requested by [[this.viewer]].
-            // TODO: for now, we begin with offset 0. This may have to be changed later (beginning somewhere in a collection)
-            this.getIncomingStillImageRepresentations(0);
-        }
-
-        // check for incoming links for the current resource
-        this.getIncomingLinks(0);
-
-
-    }
-
-    /**
-     * Gets the incoming regions for [[this.resource]].
-     *
-     * @param {number} offset the offset to be used (needed for paging). First request uses an offset of 0.
-     */
-    protected getIncomingRegions(offset: number): void {
-        this._incomingService.getIncomingRegions(this.resource.id, offset).subscribe(
-            (regions: ReadResourcesSequence) => {
-                // update ontology information
-                this.ontologyInfo.updateOntologyInformation(regions.ontologyInformation);
-
-                // Append elements of regions.resources to resource.incoming
-                Array.prototype.push.apply(this.resource.incomingRegions, regions.resources);
-
-                // prepare regions to be displayed
-                // triggers ngOnChanges of StillImageComponent
-                BeolResource.collectImagesAndRegionsForResource(this.resource);
-
-            },
-            (error: any) => {
-                this.errorMessage = <any>error;
-                this.isLoading = false;
-            }
-        );
-    }
-
-    /**
-     * Get StillImageRepresentations pointing to [[this.resource]].
-     * This method may have to called several times with an increasing offsetChange in order to get all available StillImageRepresentations.
-     *
-     * @param offset the offset to be used (needed for paging). First request uses an offset of 0.
-     * It takes the number of images returned as an argument.
-     */
-    protected getIncomingStillImageRepresentations(offset: number): void {
-        // make sure that this.resource has been initialized correctly
-        if (this.resource === undefined) {
-            return;
-        }
-
-        if (offset < 0) {
-            console.log(`offset of ${offset} is invalid`);
-            return;
-        }
-
-        this._incomingService.getStillImageRepresentationsForCompoundResource(this.resource.id, offset).subscribe(
-            (incomingImageRepresentations: ReadResourcesSequence) => {
-
-                if (incomingImageRepresentations.resources.length > 0) {
-                    // update ontology information
-                    this.ontologyInfo.updateOntologyInformation(incomingImageRepresentations.ontologyInformation);
-
-                    // set current offset
-                    this.incomingStillImageRepresentationCurrentOffset = offset;
-
-                    // TODO: implement prepending of StillImageRepresentations when moving to the left (getting previous pages)
-                    // TODO: append existing images to response and then assign response to `this.resource.incomingStillImageRepresentations`
-                    // TODO: maybe we have to support non consecutive arrays (sparse arrays)
-
-                    // append incomingImageRepresentations.resources to this.resource.incomingStillImageRepresentations
-                    Array.prototype.push.apply(this.resource.incomingStillImageRepresentations, incomingImageRepresentations.resources);
-
-                    // prepare attached image files to be displayed
-                    BeolResource.collectImagesAndRegionsForResource(this.resource);
-                }
-            },
-            (error: any) => {
-                this.errorMessage = <any>error;
-                this.isLoading = false;
-            }
-        );
-
-    }
-
-    /**
-     * Get resources pointing to [[this.resource]] with properties other than knora-api:isPartOf and knora-api:isRegionOf.
-     *
-     * @param offset the offset to be used (needed for paging). First request uses an offset of 0.
-     * It takes the number of images returned as an argument.
-     */
-    protected getIncomingLinks(offset: number): void {
-
-        this._incomingService.getIncomingLinksForResource(this.resource.id, offset).subscribe(
-            (incomingResources: ReadResourcesSequence) => {
-                // update ontology information
-                this.ontologyInfo.updateOntologyInformation(incomingResources.ontologyInformation);
-
-                // Append elements incomingResources to this.resource.incomingLinks
-                Array.prototype.push.apply(this.resource.incomingLinks, incomingResources.resources);
-            },
-            (error: any) => {
-                this.errorMessage = <any>error;
-                this.isLoading = false;
-            }
-        );
-    }
-
-    /**
-     * Gets the link value properties pointing from the incoming resource to [[this.resource]].
-     *
-     * @param {ReadResource} incomingResource the incoming resource.
-     * @returns {string} a string containing all the labels of the link value properties.
-     */
-    getIncomingPropertiesFromIncomingResource(incomingResource: ReadResource) {
-
-        const incomingProperties = [];
-
-        // collect properties, if any
-        if (incomingResource.properties !== undefined) {
-            // get property Iris (keys)
-            const propIris = Object.keys(incomingResource.properties);
-
-            // iterate over the property Iris
-            for (const propIri of propIris) {
-
-                // get the values for the current property Iri
-                const propVals: Array<ReadPropertyItem> = incomingResource.properties[propIri];
-
-                for (const propVal of propVals) {
-                    // add the property if it is a link value property pointing to [[this.resource]]
-                    if (propVal.type === KnoraConstants.LinkValue) {
-                        const linkVal = propVal as ReadLinkValue;
-
-                        if (linkVal.referredResourceIri === this.resource.id) {
-                            incomingProperties.push(propIri);
-                        }
-
-                    }
-                }
-            }
-        }
-
-        // eliminate duplicate Iris and transform to labels
-        const propLabels = incomingProperties.filter(Utils.filterOutDuplicates).map(
-            (propIri) => {
-                return this.ontologyInfo.getLabelForProperty(propIri);
-            }
-        );
-
-        // generate a string separating labels by a comma
-        return `(${propLabels.join(', ')})`;
-
-    }
 }
