@@ -1,25 +1,25 @@
-import { Component, OnInit } from '@angular/core';
-import { BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
+import { Location } from '@angular/common';
+import { Component, Inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {
-    IncomingService,
-    KnoraConstants,
-    OntologyCacheService,
-    OntologyInformation, ReadDateValue, ReadLinkValue, ReadListValue, ReadPropertyItem,
-    ReadResource,
+    Constants,
+    KnoraApiConnection,
+    ResourceClassAndPropertyDefinitions,
+    ReadValue,
     ReadTextValue,
-    ResourceService
-} from '@knora/core';
-import { Location } from '@angular/common';
-import { BeolService } from '../../services/beol.service';
-import { AppInitService } from '../../app-init.service';
+    ReadLinkValue
+} from '@dasch-swiss/dsp-js';
+import { DspApiConnectionToken, AppInitService } from '@dasch-swiss/dsp-ui';
 import { Subscription } from 'rxjs';
+import { IncomingService } from 'src/app/services/incoming.service';
+import { BeolService } from '../../services/beol.service';
+import { BeolCompoundResource, BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
 
 class CommentProps implements PropertyValues {
     commentOf: ReadLinkValue[] = [];
     text: ReadTextValue[] = [];
 
-    [index: string]: ReadPropertyItem[];
+    [index: string]: ReadValue[];
 }
 
 @Component({
@@ -30,32 +30,32 @@ class CommentProps implements PropertyValues {
 export class CommentComponent extends BeolResource {
 
     iri: string;
-    resource: ReadResource;
-    ontologyInfo: OntologyInformation;
+    resource: BeolCompoundResource;
+    ontologyInfo: ResourceClassAndPropertyDefinitions;
     incomingStillImageRepresentationCurrentOffset: number; // last offset requested for `this.resource.incomingStillImageRepresentations`
     isLoading = true;
     errorMessage: any;
     navigationSubscription: Subscription;
-    KnoraConstants = KnoraConstants;
+    dspConstants = Constants;
 
-    ontologyIri = this._appInitService.getSettings().ontologyIRI;
+    ontologyIri = this._appInitService.config['ontologyIRI'];
 
     propIris: PropIriToNameMapping = {
-        'commentOf': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#entryCommentOfValue',
-        'text': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasText',
+        'commentOf': this.ontologyIri + '/ontology/0801/beol/v2#entryCommentOfValue',
+        'text': this.ontologyIri + '/ontology/0801/beol/v2#hasText',
     };
 
     props: CommentProps;
 
-    constructor(protected _route: ActivatedRoute,
-                protected _resourceService: ResourceService,
-                protected _cacheService: OntologyCacheService,
-                protected _incomingService: IncomingService,
-                public location: Location,
-                protected _beolService: BeolService,
-                private _appInitService: AppInitService) {
+    constructor(
+        @Inject(DspApiConnectionToken) protected _dspApiConnection: KnoraApiConnection,
+        protected _route: ActivatedRoute,
+        protected _incomingService: IncomingService,
+        public location: Location,
+        protected _beolService: BeolService,
+        private _appInitService: AppInitService) {
 
-        super(_route, _resourceService, _cacheService, _incomingService, _beolService);
+        super(_dspApiConnection, _route, _incomingService, _beolService);
     }
 
     initProps() {
