@@ -1,25 +1,24 @@
-import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
-
+import { Component, Inject } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import {
-    IncomingService,
-    KnoraConstants,
-    OntologyCacheService,
-    OntologyInformation,
+    Constants,
+    KnoraApiConnection,
     ReadDateValue,
     ReadLinkValue,
     ReadListValue,
-    ReadPropertyItem,
-    ReadResource,
     ReadTextValue,
-    ReadTextValueAsString, ReadUriValue,
-    ResourceService
-} from '@knora/core';
-import { BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
+    ReadTextValueAsString,
+    ReadUriValue,
+    ReadValue,
+    ResourceClassAndPropertyDefinitions
+} from '@dasch-swiss/dsp-js';
+import { DspApiConnectionToken, AppInitService } from '@dasch-swiss/dsp-ui';
 import { Subscription } from 'rxjs';
+import { IncomingService } from 'src/app/services/incoming.service';
 import { BeolService } from '../../services/beol.service';
-import { AppInitService } from '../../app-init.service';
+import { BeolCompoundResource, BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
+import { HttpClient } from '@angular/common/http';
 
 class LetterProps implements PropertyValues {
     id: ReadTextValue[] = [];
@@ -38,7 +37,7 @@ class LetterProps implements PropertyValues {
     language: ReadTextValue[] = [];
     citation: ReadLinkValue[] = [];
 
-    [index: string]: ReadPropertyItem[];
+    [index: string]: ReadValue[];
 }
 
 @Component({
@@ -48,46 +47,49 @@ class LetterProps implements PropertyValues {
 })
 export class LeibnizLetterComponent extends BeolResource {
     iri: string;
-    resource: ReadResource;
-    ontologyInfo: OntologyInformation;
+    resource: BeolCompoundResource;
+    ontologyInfo: ResourceClassAndPropertyDefinitions;
     incomingStillImageRepresentationCurrentOffset: number; // last offset requested for `this.resource.incomingStillImageRepresentations`
     isLoading = true;
     isLoadingText = true;
     errorMessage: any;
     navigationSubscription: Subscription;
-    KnoraConstants = KnoraConstants;
+    dspConstants = Constants;
     letter;
 
+    ontologyIri = this._appInitService.config['ontologyIRI'];
+
     propIris: PropIriToNameMapping = {
-        'id': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#beolIDs',
-        'date': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#creationDate',
-        'author': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasAuthorValue',
-        'recipient': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasRecipientValue',
-        'subject': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasSubject',
-        'text': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasText',
-        'mentionedPerson': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#mentionsPersonValue',
-        'replyTo': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/leibniz/v2#isReplyToValue',
-        'location': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#location',
-        'title': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#title',
-        'letterID': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/leibniz/v2#letterID',
-        'letterURI': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#letterHasURI',
-        'language': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#letterHasLanguage',
-        'number': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#letterHasNumber',
-        'citation': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/leibniz/v2#citationValue',
+        'id': this.ontologyIri + '/ontology/0801/beol/v2#beolIDs',
+        'date': this.ontologyIri + '/ontology/0801/beol/v2#creationDate',
+        'author': this.ontologyIri + '/ontology/0801/beol/v2#hasAuthorValue',
+        'recipient': this.ontologyIri + '/ontology/0801/beol/v2#hasRecipientValue',
+        'subject': this.ontologyIri + '/ontology/0801/beol/v2#hasSubject',
+        'text': this.ontologyIri + '/ontology/0801/beol/v2#hasText',
+        'mentionedPerson': this.ontologyIri + '/ontology/0801/beol/v2#mentionsPersonValue',
+        'replyTo': this.ontologyIri + '/ontology/0801/leibniz/v2#isReplyToValue',
+        'location': this.ontologyIri + '/ontology/0801/beol/v2#location',
+        'title': this.ontologyIri + '/ontology/0801/beol/v2#title',
+        'letterID': this.ontologyIri + '/ontology/0801/leibniz/v2#letterID',
+        'letterURI': this.ontologyIri + '/ontology/0801/beol/v2#letterHasURI',
+        'language': this.ontologyIri + '/ontology/0801/beol/v2#letterHasLanguage',
+        'number': this.ontologyIri + '/ontology/0801/beol/v2#letterHasNumber',
+        'citation': this.ontologyIri + '/ontology/0801/leibniz/v2#citationValue',
     };
 
     props: LetterProps;
 
-    constructor (protected _route: ActivatedRoute,
-        protected _resourceService: ResourceService,
-        protected _cacheService: OntologyCacheService,
+    constructor(
+        @Inject(DspApiConnectionToken) protected _dspApiConnection: KnoraApiConnection,
+        protected _route: ActivatedRoute,
         protected _incomingService: IncomingService,
         public location: Location,
         protected _beolService: BeolService,
-        private _appInitService: AppInitService
+        private _appInitService: AppInitService,
+        private _http: HttpClient
     ) {
 
-        super(_route, _resourceService, _cacheService, _incomingService, _beolService);
+        super(_dspApiConnection, _route, _incomingService, _beolService);
 
     }
 
@@ -100,7 +102,7 @@ export class LeibnizLetterComponent extends BeolResource {
         this.props = props;
 
         // get the id from the route leibnizLetter/:id e.g. l386
-        this.getLeibnizLetterText(this.props.letterID[0].getContent());
+        this.getLeibnizLetterText(this.props.letterID[0].strval);
     }
 
 
@@ -108,25 +110,26 @@ export class LeibnizLetterComponent extends BeolResource {
         // use a proxy url as described here:
         // https://stackoverflow.com/questions/43871637/no-access-control-allow-origin-header-is-present-on-the-requested-resource-whe
         // const proxyurl = 'https://cors-anywhere.herokuapp.com//';
-        const basePath = this._appInitService.getSettings().leibnizApi + 'select?sort=type+asc&q=id%3A';
+        const basePath = this._appInitService.config['leibnizApi'] + 'select?sort=type+asc&q=id%3A';
         const basePathOR = '+OR+(doc_id%3A';
         const basePathAnd = '+AND+type%3Avariante)&rows=9999&wt=json';
         const apiUrl = basePath + filename + basePathOR + filename + basePathAnd; // site that doesn’t send Access-Control-*
 
-        fetch( apiUrl) // could be proxyurl + apiURL as https://cors-anywhere.herokuapp.com/https://example.com
-            .then(response => response.json())
-            .then(contents => {
+        this._http.get(apiUrl) // could be proxyurl + apiURL as https://cors-anywhere.herokuapp.com/https://example.com
+            //.then(response => response.json())
+            .subscribe(contents => {
 
                 this.getLeibnizLetterBody(contents);
                 this.isLoadingText = false;
-            })
-            .catch(e => console.log('Can’t access ' + apiUrl + ' response. Blocked by browser?', e));
+            },
+                err => console.log('Can’t access ' + apiUrl + ' response. Blocked by browser?', err)
+            );
     }
 
 
     private getLeibnizImages(element) {
         const proxyurl = 'https://cors-anywhere.herokuapp.com/';
-        const basePath = this._appInitService.getSettings().leibnizApi + 'select?q=id%3A';
+        const basePath = this._appInitService.config['leibnizApi'] + 'select?q=id%3A';
         const basePathTail = '&rows=9999&wt=json';
 
         const imgs = element.getElementsByTagName('span');
@@ -135,14 +138,14 @@ export class LeibnizLetterComponent extends BeolResource {
             if (image.getAttribute('class') === 'reference -image') {
                 const filename = image.getAttribute('data-id');
                 const apiUrl = basePath + filename + basePathTail; // get the svg element
-                fetch(proxyurl + apiUrl) // https://cors-anywhere.herokuapp.com/https://example.com
-                    .then(response => response.json())
-                    .then(contents => {
+                this._http.get(proxyurl + apiUrl) // https://cors-anywhere.herokuapp.com/https://example.com
+                    //.then(response => response.json())
+                    .subscribe(contents => {
                         const svgElement = this.getLeibnizImageSVG(contents);
                         image.replaceWith(svgElement);
                     });
 
-                console.log(this._appInitService.getSettings().appURL);
+                // console.log(this._appInitService.config['appURL']);
             }
 
         }
