@@ -3,39 +3,46 @@ import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { MaterialModule } from '../../material-module';
 import { HttpClientModule } from '@angular/common/http';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { of } from 'rxjs';
 import { NewtonLetterComponent } from './newton-letter.component';
-import { KuiActionModule } from '@knora/action';
-import { KuiCoreConfig, KuiCoreConfigToken, OntologyCacheService } from '@knora/core';
 import { ReadTextValueAsHtmlComponent } from '../../properties/read-text-value-as-html/read-text-value-as-html.component';
 import { ReadListValueComponent } from '../../properties/read-list-value/read-list-value.component';
 import { MathJaxDirective } from '../../directives/mathjax.directive';
-import { KuiViewerModule } from '@knora/viewer';
 import { ReadTextValueComponent } from '../../properties/read-text-value/read-text-value.component';
 import { NewtonProjectDirective } from '../../directives/newton-project.directive';
 import { SanitizeHtmlPipe } from '../../pipes/sanitize-html.pipe';
-import { AppInitService } from '../../app-init.service';
+import { AppInitService, DspActionModule, DspApiConnectionToken } from '@dasch-swiss/dsp-ui';
+import { ReadResource, ReadResourceSequence, ReadTextValueAsString, ResourcesEndpointV2, SearchEndpointV2 } from '@dasch-swiss/dsp-js';
 
 describe('NewtonLetterComponent', () => {
     let component: NewtonLetterComponent;
     let fixture: ComponentFixture<NewtonLetterComponent>;
 
-    let appInitService: AppInitService;
-
     const id = 'http://rdfh.ch/0801/7ZvL2A5PQ9C4eAmr-n26gw';
 
     beforeEach(async(() => {
-        const appInitServiceSpy = jasmine.createSpyObj('AppInitService', ['getSettings']);
+
+        const dspConnectionSpy = {
+            v2: {
+                res: jasmine.createSpyObj('res', ['getResource']),
+                search: jasmine.createSpyObj('search', ['doExtendedSearch'])
+            }
+        };
+
+        const appInitServiceMock = {
+            config: {
+                ontologyIRI: 'http://0.0.0.0:3333'
+            }
+        };
 
         TestBed.configureTestingModule({
             imports: [
-                KuiActionModule,
-                KuiViewerModule,
                 MaterialModule,
                 RouterTestingModule,
                 HttpClientModule,
-                HttpClientTestingModule
+                HttpClientTestingModule,
+                DspActionModule
             ],
             declarations: [
                 NewtonLetterComponent,
@@ -47,36 +54,67 @@ describe('NewtonLetterComponent', () => {
                 SanitizeHtmlPipe
             ],
             providers: [
-                OntologyCacheService,
                 {
                     provide: ActivatedRoute,
-                    useValue: { paramMap: of({
+                    useValue: {
+                        paramMap: of({
                             get: () => {
                                 return id;
                             }
-                        })}
+                        })
+                    }
                 },
-                { provide: KuiCoreConfigToken, useValue: KuiCoreConfig },
-                { provide: AppInitService, useValue: appInitServiceSpy }
+                { provide: AppInitService, useValue: appInitServiceMock },
+                { provide: DspApiConnectionToken, useValue: dspConnectionSpy }
             ]
         })
             .compileComponents();
 
-        appInitServiceSpy.getSettings.and.returnValue({ontologyIRI: 'http://0.0.0.0:3333'});
 
-        appInitService = TestBed.get(AppInitService);
 
     }));
 
     beforeEach(() => {
+        const dspServiceSpy = TestBed.inject(DspApiConnectionToken);
+
+        const res = new ReadResource();
+        const letterIdVal = new ReadTextValueAsString();
+        letterIdVal.strval = '1';
+        letterIdVal.property = 'http://0.0.0.0:3333/ontology/0801/newton/v2#newtonProjectID';
+
+        res.properties = {
+            'http://0.0.0.0:3333/ontology/0801/leibniz/v2#letterID': [letterIdVal]
+        };
+
+        (dspServiceSpy.v2.res as jasmine.SpyObj<ResourcesEndpointV2>).getResource.and.returnValue(of(res));
+
+        (dspServiceSpy.v2.search as jasmine.SpyObj<SearchEndpointV2>).doExtendedSearch.and.returnValue(of(new ReadResourceSequence([])));
+
+        // TODO: use Angular's HTTP client
+        const fetchSpy = spyOn(window, 'fetch').and.callFake(
+            path => {
+                return Promise.resolve(new Response(''));
+            }
+        );
+
         fixture = TestBed.createComponent(NewtonLetterComponent);
         component = fixture.componentInstance;
         fixture.detectChanges();
     });
 
     it('should create', () => {
-        expect(component).toBeTruthy();
+        const httpTestingController = TestBed.inject(HttpTestingController);
 
-        expect(appInitService.getSettings).toHaveBeenCalled();
+        // TODO: use Angular's HTTP client
+        /*httpTestingController
+            .expectOne('https://cors-anywhere.herokuapp.com/http://www.newtonproject.ox.ac.uk/view/texts/normalized/1')
+            .flush('');*/
+
+        expect(component).toBeTruthy();
+    });
+
+    afterEach(() => {
+        const httpTestingController = TestBed.inject(HttpTestingController);
+        httpTestingController.verify();
     });
 });
