@@ -1,21 +1,19 @@
-import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
+import { Component, Inject } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import {
-    IncomingService,
-    KnoraConstants,
-    OntologyCacheService,
-    OntologyInformation,
+    Constants,
+    KnoraApiConnection,
     ReadDateValue,
-    ReadPropertyItem,
-    ReadResource,
     ReadTextValue,
-    ResourceService,
-} from '@knora/core';
-import { BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
+    ReadValue,
+    ResourceClassAndPropertyDefinitions
+} from '@dasch-swiss/dsp-js';
+import { DspApiConnectionToken, AppInitService } from '@dasch-swiss/dsp-ui';
 import { Subscription } from 'rxjs';
+import { IncomingService } from 'src/app/services/incoming.service';
 import { BeolService } from '../../services/beol.service';
-import { AppInitService } from '../../app-init.service';
+import { BeolCompoundResource, BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
 
 class PersonProps implements PropertyValues {
     comment: ReadTextValue[] = [];
@@ -25,12 +23,12 @@ class PersonProps implements PropertyValues {
     deathDate: ReadDateValue[] = [];
     deathPlace: ReadTextValue[] = [];
     dictionary: ReadTextValue[] = [];
-    givenName:  ReadTextValue[] = [];
-    familyName:  ReadTextValue[] = [];
-    title:  ReadTextValue[] = [];
+    givenName: ReadTextValue[] = [];
+    familyName: ReadTextValue[] = [];
+    title: ReadTextValue[] = [];
     IAF: ReadTextValue[] = [];
     mentioned: ReadTextValue[] = [];
-    [index: string]: ReadPropertyItem[];
+    [index: string]: ReadValue[];
 }
 
 @Component({
@@ -41,42 +39,44 @@ class PersonProps implements PropertyValues {
 export class PersonComponent extends BeolResource {
 
     iri: string;
-    resource: ReadResource;
-    ontologyInfo: OntologyInformation;
+    resource: BeolCompoundResource;
+    ontologyInfo: ResourceClassAndPropertyDefinitions;
     incomingStillImageRepresentationCurrentOffset: number; // last offset requested for `this.resource.incomingStillImageRepresentations`
     isLoading = true;
     errorMessage: any;
     navigationSubscription: Subscription;
-    KnoraConstants = KnoraConstants;
+    dspConstants = Constants;
+
+    ontologyIri = this._appInitService.config['ontologyIRI'];
 
     propIris: PropIriToNameMapping = {
-        'id': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#beolIDs',
-        'comment': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#comment',
-        'alternative': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasAlternativeName',
-        'birthDate': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasBirthDate',
-        'birthPlace': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasBirthPlace',
-        'deathDate': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasDeathDate',
-        'deathPlace': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasDeathPlace',
-        'dictionary': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasDictionaryEntries',
-        'familyName': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasFamilyName',
-        'givenName': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasGivenName',
-        'title': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#personHasTitle',
-        'IAF': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasIAFIdentifier',
-        'mentioned': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#mentionedIn',
+        'id': this.ontologyIri + '/ontology/0801/beol/v2#beolIDs',
+        'comment': this.ontologyIri + '/ontology/0801/beol/v2#comment',
+        'alternative': this.ontologyIri + '/ontology/0801/beol/v2#hasAlternativeName',
+        'birthDate': this.ontologyIri + '/ontology/0801/beol/v2#hasBirthDate',
+        'birthPlace': this.ontologyIri + '/ontology/0801/beol/v2#hasBirthPlace',
+        'deathDate': this.ontologyIri + '/ontology/0801/beol/v2#hasDeathDate',
+        'deathPlace': this.ontologyIri + '/ontology/0801/beol/v2#hasDeathPlace',
+        'dictionary': this.ontologyIri + '/ontology/0801/beol/v2#hasDictionaryEntries',
+        'familyName': this.ontologyIri + '/ontology/0801/beol/v2#hasFamilyName',
+        'givenName': this.ontologyIri + '/ontology/0801/beol/v2#hasGivenName',
+        'title': this.ontologyIri + '/ontology/0801/beol/v2#personHasTitle',
+        'IAF': this.ontologyIri + '/ontology/0801/beol/v2#hasIAFIdentifier',
+        'mentioned': this.ontologyIri + '/ontology/0801/beol/v2#mentionedIn',
     };
 
     props: PersonProps;
 
-    constructor(protected _route: ActivatedRoute,
-                protected _resourceService: ResourceService,
-                protected _cacheService: OntologyCacheService,
-                protected _incomingService: IncomingService,
-                public location: Location,
-                protected _beolService: BeolService,
-                private _appInitService: AppInitService
+    constructor(
+        @Inject(DspApiConnectionToken) protected _dspApiConnection: KnoraApiConnection,
+        protected _route: ActivatedRoute,
+        protected _incomingService: IncomingService,
+        public location: Location,
+        protected _beolService: BeolService,
+        private _appInitService: AppInitService
     ) {
 
-        super(_route, _resourceService, _cacheService, _incomingService, _beolService);
+        super(_dspApiConnection, _route, _incomingService, _beolService);
 
     }
 
@@ -88,6 +88,7 @@ export class PersonComponent extends BeolResource {
 
         this.props = props;
     }
+
     showIncomingRes(resIri, resType) {
         this._beolService.routeByResourceType(resType, resIri);
     }
