@@ -1,28 +1,26 @@
-import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
+import { Component, Inject } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import {
-    IncomingService,
-    KnoraConstants,
-    OntologyCacheService,
-    OntologyInformation,
+    Constants,
+    KnoraApiConnection,
     ReadLinkValue,
-    ReadPropertyItem,
-    ReadResource,
     ReadTextValue,
-    ResourceService,
-} from '@knora/core';
-import { BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
+    ReadValue,
+    ResourceClassAndPropertyDefinitions
+} from '@dasch-swiss/dsp-js';
+import { DspApiConnectionToken, AppInitService } from '@dasch-swiss/dsp-ui';
 import { Subscription } from 'rxjs';
+import { IncomingService } from 'src/app/services/incoming.service';
 import { BeolService } from '../../services/beol.service';
-import { AppInitService } from '../../app-init.service';
+import { BeolCompoundResource, BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
 
 class PublishedLetterProps implements PropertyValues {
     publishedLetterNumber: ReadTextValue[] = [];
     startPage: ReadTextValue[] = [];
     endPage: ReadTextValue[] = [];
     isPublishedIn: ReadLinkValue[] = [];
-    [index: string]: ReadPropertyItem[];
+    [index: string]: ReadValue[];
 }
 
 @Component({
@@ -33,34 +31,34 @@ class PublishedLetterProps implements PropertyValues {
 export class PublishedLetterComponent extends BeolResource {
 
     iri: string;
-    resource: ReadResource;
-    ontologyInfo: OntologyInformation;
+    resource: BeolCompoundResource;
+    ontologyInfo: ResourceClassAndPropertyDefinitions;
     incomingStillImageRepresentationCurrentOffset: number; // last offset requested for `this.resource.incomingStillImageRepresentations`
     isLoading = true;
     errorMessage: any;
     navigationSubscription: Subscription;
-    KnoraConstants = KnoraConstants;
+    dspConstants = Constants;
 
     propIris: PropIriToNameMapping = {
-        'id': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#beolIDs',
-        'isPublishedIn': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/biblio/v2#isPublishedInValue',
-        'publishedLetterNumber': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/biblio/v2#publishedLetterNumber',
-        'startPage': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/biblio/v2#publishedLetterStartPage',
-        'endPage': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/biblio/v2#publishedLetterEndPage'
+        'id': this._appInitService.config['ontologyIRI'] + '/ontology/0801/beol/v2#beolIDs',
+        'isPublishedIn': this._appInitService.config['ontologyIRI'] + '/ontology/0801/biblio/v2#isPublishedInValue',
+        'publishedLetterNumber': this._appInitService.config['ontologyIRI'] + '/ontology/0801/biblio/v2#publishedLetterNumber',
+        'startPage': this._appInitService.config['ontologyIRI'] + '/ontology/0801/biblio/v2#publishedLetterStartPage',
+        'endPage': this._appInitService.config['ontologyIRI'] + '/ontology/0801/biblio/v2#publishedLetterEndPage'
     };
 
     props: PublishedLetterProps;
 
-    constructor(protected _route: ActivatedRoute,
-                protected _resourceService: ResourceService,
-                protected _cacheService: OntologyCacheService,
-                protected _incomingService: IncomingService,
-                public location: Location,
-                protected _beolService: BeolService,
-                private _appInitService: AppInitService
+    constructor(
+        @Inject(DspApiConnectionToken) protected _dspApiConnection: KnoraApiConnection,
+        protected _route: ActivatedRoute,
+        protected _incomingService: IncomingService,
+        public location: Location,
+        protected _beolService: BeolService,
+        private _appInitService: AppInitService
     ) {
 
-        super(_route, _resourceService, _cacheService, _incomingService, _beolService);
+        super(_dspApiConnection, _route, _incomingService, _beolService);
 
     }
 
