@@ -1,32 +1,30 @@
-import { Component } from '@angular/core';
-import { BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
-import {
-    IncomingService,
-    KnoraConstants,
-    OntologyCacheService,
-    OntologyInformation,
-    ReadIntegerValue,
-    ReadLinkValue,
-    ReadPropertyItem,
-    ReadResource,
-    ReadResourcesSequence,
-    ReadTextValue,
-    ResourceService,
-    SearchService
-} from '@knora/core';
 import { Location } from '@angular/common';
+import { Component, Inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { BeolService } from '../../services/beol.service';
+import {
+    Constants,
+    KnoraApiConnection,
+    ReadIntValue,
+    ReadLinkValue,
+    ReadResource,
+    ReadResourceSequence,
+    ReadTextValue,
+    ReadValue,
+    ResourceClassAndPropertyDefinitions
+} from '@dasch-swiss/dsp-js';
+import { DspApiConnectionToken, AppInitService } from '@dasch-swiss/dsp-ui';
 import { Subscription } from 'rxjs';
-import { AppInitService } from '../../app-init.service';
+import { IncomingService } from 'src/app/services/incoming.service';
+import { BeolService } from '../../services/beol.service';
+import { BeolCompoundResource, BeolResource, PropertyValues, PropIriToNameMapping } from '../beol-resource';
 
 class TranscriptionProps implements PropertyValues {
     text: ReadTextValue[] = [];
-    layer: ReadIntegerValue[] = [];
+    layer: ReadIntValue[] = [];
     transcriptionOf: ReadLinkValue[] = [];
     belongsToRegion: ReadLinkValue[] = [];
 
-    [index: string]: ReadPropertyItem[];
+    [index: string]: ReadValue[];
 }
 
 @Component({
@@ -37,36 +35,35 @@ class TranscriptionProps implements PropertyValues {
 export class TranscriptionComponent extends BeolResource {
 
     iri: string;
-    resource: ReadResource;
-    ontologyInfo: OntologyInformation;
+    resource: BeolCompoundResource;
+    ontologyInfo: ResourceClassAndPropertyDefinitions;
     incomingStillImageRepresentationCurrentOffset: number; // last offset requested for `this.resource.incomingStillImageRepresentations`
     isLoading = true;
     errorMessage: any;
     navigationSubscription: Subscription;
-    KnoraConstants = KnoraConstants;
+    dspConstants = Constants;
 
     otherLayers: ReadResource[] = [];
 
     propIris: PropIriToNameMapping = {
-        'text': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#hasText',
-        'layer': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#layer',
-        'transcriptionOf': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#transcriptionOfValue',
-        'belongsToRegion': this._appInitService.getSettings().ontologyIRI + '/ontology/0801/beol/v2#belongsToRegionValue'
+        'text': this._appInitService.config['ontologyIRI'] + '/ontology/0801/beol/v2#hasText',
+        'layer': this._appInitService.config['ontologyIRI'] + '/ontology/0801/beol/v2#layer',
+        'transcriptionOf': this._appInitService.config['ontologyIRI'] + '/ontology/0801/beol/v2#transcriptionOfValue',
+        'belongsToRegion': this._appInitService.config['ontologyIRI'] + '/ontology/0801/beol/v2#belongsToRegionValue'
 
     };
 
     props: TranscriptionProps;
 
-    constructor(protected _route: ActivatedRoute,
-                protected _resourceService: ResourceService,
-                protected _cacheService: OntologyCacheService,
-                protected _incomingService: IncomingService,
-                private _searchService: SearchService,
-                public location: Location,
-                protected _beolService: BeolService,
-                private _appInitService: AppInitService) {
+    constructor(
+        @Inject(DspApiConnectionToken) protected _dspApiConnection: KnoraApiConnection,
+        protected _route: ActivatedRoute,
+        protected _incomingService: IncomingService,
+        public location: Location,
+        protected _beolService: BeolService,
+        private _appInitService: AppInitService) {
 
-        super(_route, _resourceService, _cacheService, _incomingService, _beolService);
+        super(_dspApiConnection, _route, _incomingService, _beolService);
     }
 
     initProps() {
@@ -88,19 +85,18 @@ export class TranscriptionComponent extends BeolResource {
         }
 
         const otherLayersForManEntry =
-            this._beolService.getTranscriptionsForManuscriptEntry(this.props.transcriptionOf[0].referredResourceIri, this.props.layer[0].integer, true);
+            this._beolService.getTranscriptionsForManuscriptEntry(
+                this.props.transcriptionOf[0].linkedResourceIri,
+                this.props.layer[0].int, true
+            );
 
-        this._searchService.doExtendedSearchReadResourceSequence(otherLayersForManEntry).subscribe(
-            (otherLayers: ReadResourcesSequence) => {
-                if (otherLayers.numberOfResources > 0) {
+        this._dspApiConnection.v2.search.doExtendedSearch(otherLayersForManEntry).subscribe(
+            (otherLayers: ReadResourceSequence) => {
+                if (otherLayers.resources.length > 0) {
                     this.otherLayers = otherLayers.resources;
                 }
             }
         );
-    }
-
-    showIncomingRes(resIri, resType) {
-        this._beolService.routeByResourceType(resType, resIri);
     }
 
     goToResource(resType: string, resIri: string) {
